+++ R/add_dockerfiles_renv.R
@@ -136,11 +136,10 @@ add_dockerfile_with_renv_ <- function(
     
   }
   
-  
+  if (!single_file){
   my_dock$COPY(basename(lockfile), "renv.lock")
-  
   my_dock$RUN("R -e 'options(renv.config.pak.enabled = FALSE);renv::restore()'")
-  
+  }
   # we use an already built tar.gz file
   my_dock$COPY(
     from =
@@ -237,8 +236,13 @@ add_dockerfile_with_renv_ <- function(
 #' See available distributions at https://hub.docker.com/r/rstudio/r-base/.
 #' @param document boolean. If TRUE (by default), DESCRIPTION file is updated using [attachment::att_amend_desc()] before creating the renv.lock file
 #' @param dockerfile_cmd What is the CMD to add to the Dockerfile. If NULL, the default,
-#' the CMD will be `R -e "options('shiny.port'={port},shiny.host='{host}');library({appname});{appname}::run_app()\`.
+#' the CMD will be `R -e "options('shiny.port'={port},shiny.host='{host}',golem.app.prod = {set_golem.app.prod});library({appname});{appname}::run_app()\`.
 #' @param user Name of the user to specify in the Dockerfile with the USER instruction. Default is `rstudio`, if set to `NULL` no the user from the FROM image is used.
+#' @param single_file boolean.  
+#'   If `TRUE` (by default), generate a single multi-stage Dockerfile .  
+#'   If `FALSE`, produce two distinct Dockerfiles to be run sequentially
+#'   for the build and production phases.
+#' @param set_golem.app.prod boolean If `TRUE` (by default) set options(golem.app.prod = TRUE) in dockerfile_cmd. 
 #' @param ... Other arguments to pass to [renv::snapshot()].
 #' @inheritParams add_dockerfile
 #' @rdname dockerfiles
@@ -261,7 +265,8 @@ add_dockerfile_with_renv <- function(
   update_tar_gz = TRUE,
   dockerfile_cmd = NULL,
   user = "rstudio",
-  single_file = FALSE,
+  single_file = TRUE,
+  set_golem.app.prod = TRUE,
   ...
 ) {
   base_dock <- add_dockerfile_with_renv_(
@@ -288,9 +293,10 @@ add_dockerfile_with_renv <- function(
   }
   if (is.null(dockerfile_cmd)) {
     dockerfile_cmd <- sprintf(
-      "R -e \"options('shiny.port'=%s,shiny.host='%s');library(%3$s);%3$s::run_app()\"",
+      "R -e \"options('shiny.port'=%s,shiny.host='%s',golem.app.prod=%s);library(%3$s);%3$s::run_app()\"",
       port,
       host,
+      set_golem.app.prod,
       get_golem_name(
         pkg = source_folder
       )
@@ -306,9 +312,8 @@ add_dockerfile_with_renv <- function(
   
   if (!single_file){
   out <- sprintf(
-    "
-    #cd to/your/output/dir
-    docker build -f Dockerfile_base --progress=plain -t %s .
+"# use cd to moove to the folder containing the Dockerfile
+docker build -f Dockerfile_base --progress=plain -t %s .
 docker build -f Dockerfile --progress=plain -t %s .
 docker run -p %s:%s %s
 # then go to 127.0.0.1:%s",
@@ -341,9 +346,8 @@ docker run -p %s:%s %s
     
     
     out <- sprintf(
-      "
-      #cd to/your/output/dir
-      docker build -f Dockerfile --target=final --progress=plain -t %s .
+"# use cd to moove to the folder containing the Dockerfile
+docker build -f Dockerfile --target=final --progress=plain -t %s .
 docker run -p %s:%s %s
 # then go to 127.0.0.1:%s",
       tolower(paste0(
@@ -363,19 +367,8 @@ docker run -p %s:%s %s
       port
     )
     
-    
-    
-    
-    
-    
-    
   }
   
-  
-  
-  
-  
-
   cat(out, file = file.path(output_dir, "README"))
 
   open_or_go_to(
@@ -403,7 +396,8 @@ add_dockerfile_with_renv_shinyproxy <- function(
   document = TRUE,
   update_tar_gz = TRUE,
   user = "rstudio",
-  single_file = FALSE,
+  single_file = TRUE,
+  set_golem.app.prod = TRUE,
   ...
 ) {
   add_dockerfile_with_renv(
@@ -424,6 +418,7 @@ add_dockerfile_with_renv_shinyproxy <- function(
     document = document,
     user = user,
     single_file = single_file,
+    set_golem.app.prod = set_golem.app.prod,
     dockerfile_cmd = sprintf(
       "R -e \"options('shiny.port'=3838,shiny.host='0.0.0.0');library(%1$s);%1$s::run_app()\"",
       get_golem_name(
@@ -453,7 +448,8 @@ add_dockerfile_with_renv_heroku <- function(
   document = TRUE,
   user = "rstudio",
   update_tar_gz = TRUE,
-  single_file = FALSE,
+  single_file = TRUE,
+  set_golem.app.prod = TRUE,
   ...
 ) {
   add_dockerfile_with_renv(
@@ -474,6 +470,7 @@ add_dockerfile_with_renv_heroku <- function(
     document = document,
     user = user,
     single_file = single_file,
+    set_golem.app.prod = set_golem.app.prod,
     dockerfile_cmd = sprintf(
       "R -e \"options('shiny.port'=$PORT,shiny.host='0.0.0.0');library(%1$s);%1$s::run_app()\"",
       get_golem_name(